chat/requirements.txt
docx==0.2.4
langchain==0.0.319
openpyxl==3.1.2
pgvector==0.2.3
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
pydantic-core==2.10.1
pypdf==3.16.4
sentence-transformers==2.2.2
streamlit==1.28.0
unstructured==0.10.27
flask==3.0.0
python-dotenv==1.0.0
pyjwt==2.3.0
cryptography==41.0.4